fixup! Query service uses template_search instead of search
